Refactor to seperate files with import and export
